requirements.txt
# Core Flask
Flask==2.3.3
Werkzeug==2.3.7

# Extensions
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0

# Database
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.0.0

# Server
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
eventlet==0.33.3

# Validation & Security
email-validator==2.1.0
phonenumbers==8.13.26
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# AI & ML
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
